Move `Navigation` component into the `App` folder
